Also rename vm_test to vm_tests, for consistency with typechecker_tests
